Restructure the package to standard Dart lib/src split
